latitude and longitude change long to double

+++ src/main/java/com/example/demo/dto/AddressModel.java
@@ -28,8 +28,8 @@ public class AddressModel {
 	private String state;
 	private String country;
 
-	private long latitude;
-	private long longitude;
+	private double latitude;
+	private double longitude;
 
 
 
@@ -73,23 +73,23 @@ public class AddressModel {
 		this.country = country;
 	}
 
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
-	public AddressModel(String deploymentId, String streetName, String city, String pinCode, String state, String country,long longitude,long latitude) {
+	public AddressModel(String deploymentId, String streetName, String city, String pinCode, String state, String country, double longitude, double latitude) {
 		super();
 		this.deploymentId = deploymentId;
 		this.streetName = streetName;
